Unified loader & first step for "index" standardization
Roadmap: diagnosed or user-defined "index" that is then used for outer
products / by=...
For now the index is not really used, and standard names are used for
grouping.
Eventually, the FastIamDataFrame class should be used without looking
back at IamDataFrame, so that we can go our own way for the specific
need of rime. (e.g. we don't always need a time dimension).
The only thing we need is idenfying the dimensions for warming_level,
year and for doing outer products vs pooling.