major update for rime-run-timeseries: bug fixes + simplified inputs
This version is less opinionated about how the input dataset should look
like. It leaves the user specify the index on which group operations
are made, and drop/pool any dimension that should not appear in the
result.

--average-scenario, --average-model --make-equiprobable-mdoels are
dropped in favour of more generic --average midyear scenario --average model
or --equiprobable-dim model kind of options