folder structure reorganization (preprocessing) + better option-passing
1. Climate_impact_explorer is now the root folder under which binned data are
stored, and a child isimip2 and isimip3 is automatically added based on
the simulation round (ISIMIP2b or ISIMIP3b).
2. A further sub-folder is created based on settings that affect the
   calibration data, and warming levels and binned impact data are
stored underneath

The --simulation-round option (along with --projection-baseline) are thus
sufficient to switch between the various isimip version without editing
a new config file.

Additionally, the "selection" option was removed since it can be handled
at run time. Consequently, the experiment name currently only contains
the running mean length.

Note this structure is still insatisfactory and will be simplified in the
future.